Meraki admission component refactored
